+++ meta_package_manager/managers/apt.py
@@ -19,7 +19,8 @@ import re
 
 from click_extra.platform import LINUX
 
-from ..base import Package, PackageManager, no_extended_search
+from ..base import Package, PackageManager
+from ..capabilities import search_capabilities
 
 
 class APT(PackageManager):
@@ -274,7 +275,7 @@ class APT_Mint(APT):
         1.6.11
     """
 
-    @no_extended_search
+    @search_capabilities(extended_support=False)
     def search(self, query, extended, exact):
         """Fetch matching packages.