+++ meta_package_manager/managers/composer.py
@@ -20,7 +20,8 @@ import re
 
 from click_extra.platform import LINUX, MACOS, WINDOWS
 
-from ..base import Package, PackageManager, no_exact_search
+from ..base import Package, PackageManager
+from ..capabilities import search_capabilities
 
 
 class Composer(PackageManager):
@@ -119,7 +120,7 @@ class Composer(PackageManager):
                     latest_version=package["latest"],
                 )
 
-    @no_exact_search
+    @search_capabilities(exact_support=False)
     def search(self, query, extended, exact):
         """Fetch matching packages.