+++ meta_package_manager/managers/mas.py
@@ -19,7 +19,8 @@ import re
 
 from click_extra.platform import MACOS
 
-from ..base import Package, PackageManager, no_exact_search, no_extended_search
+from ..base import Package, PackageManager
+from ..capabilities import search_capabilities
 
 
 class MAS(PackageManager):
@@ -113,8 +114,7 @@ class MAS(PackageManager):
                 latest_version=latest_version,
             )
 
-    @no_exact_search
-    @no_extended_search
+    @search_capabilities(extended_support=False, exact_support=False)
     def search(self, query, extended, exact):
         """Fetch matching packages.