+++ meta_package_manager/managers/opkg.py
@@ -19,7 +19,8 @@ import re
 
 from click_extra.platform import LINUX
 
-from ..base import Package, PackageManager, no_exact_search, no_extended_search
+from ..base import Package, PackageManager
+from ..capabilities import search_capabilities
 
 
 class OPKG(PackageManager):
@@ -98,8 +99,7 @@ class OPKG(PackageManager):
                     installed_version=installed_version,
                 )
 
-    @no_exact_search
-    @no_extended_search
+    @search_capabilities(extended_support=False, exact_support=False)
     def search(self, query, extended, exact):
         """Fetch matching packages.